+++ src/core/prelude/i18n/const.ts
@@ -1,0 +1,3 @@
+export * from '@v4fire/core/core/prelude/i18n/const';
+
+export const COOKIE_STORAGE_NAME = 'v4s';