+++ src/form/b-select-date/b-select-date.i18n/ru.js
@@ -1,0 +1,24 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+module.exports = {
+	'b-select-date': {
+		January: 'Январь',
+		February: 'Февраль',
+		March: 'Март',
+		April: 'Апрель',
+		May: 'Май',
+		June: 'Июнь',
+		July: 'Июль',
+		August: 'Август',
+		September: 'Сентябрь',
+		October: 'Октябрь',
+		November: 'Ноябрь',
+		December: 'Декабрь'
+	}
+};