+++ src/super/i-static-page/i-static-page.ts
@@ -133,14 +133,14 @@ export default abstract class iStaticPage extends iPage {
 	/**
 	 * System locale
 	 */
-	get locale(): string {
-		return this.field.get<string>('localeStore')!;
+	get locale(): Language {
+		return this.field.get<Language>('localeStore')!;
 	}
 
 	/**
 	 * Sets a new system locale
 	 */
-	set locale(value: string) {
+	set locale(value: Language) {
 		this.field.set('localeStore', value);
 
 		try {
@@ -318,7 +318,7 @@ export default abstract class iStaticPage extends iPage {
 	 * @param locale
 	 */
 	@watch(['localeStore', 'globalEmitter:i18n.setLocale'])
-	protected syncLocaleWatcher(locale: string): void {
+	protected syncLocaleWatcher(locale: Language): void {
 		if (this.locale === locale) {
 			return;
 		}